Added gamepad support (#1483)
* Added default gamepad bindings. Hidden behind a tag.
* Added gamepad tester ui

I'm not sure about that bindings being disabled by default and hidden
behind a tag. That makes it so that whenever someone wants to try out a
gamepad they need to figure out how to enable it first. If you don't
have a gamepad connected these bindings will do nothing. Maybe we should
have it be active by default and have a tag to disable it in case
someone wants to use their own custom gamepad bindings? This code is
flexible enough that you can override every single key in a Talon action
so you shouldn't really need to do anything else, but you know.
What are peoples take on this? Enabled or disabled by default?

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>