Added phrase enders list (#1544)
Instead of just `"over"`, adds a list of symbols that can be used to
end/terminate a phrase. Also supports values for these phrases so you
can end your phrase with a symbol that will be inserted. eg `"say hello
there slap"` => `hello there\n`

This implementation is quite simple with the drawback that there is no
synchronization between the spoken forms in this list and existing voice
commands for adding these symbols.

Slack reference
https://talonvoice.slack.com/archives/C9MHQ4AGP/p1723643676364569

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>