Add more per-platform keys in talon-lists. (#1574)
All platforms:
- keypad_enter
- insert (inadvertently removed in #1554)
- Allow dictating keypad keys directly without "press"

Mac:
- "function"/"globe" for globe/fn modifier (increasingly used by the OS)
- "return" for return (as printed on Apple keyboards)
- "enter" for [keypad] enter

The only one I'm a bit worried about is the last one. It is correct per
Apple's definitions but it's possible some users are used to saying
"enter" (though it's not the correct name for the key) and will get
unexpected behavior; in most Mac apps I've tried, Return and Enter
behave identically.

Co-authored-by: Jeff Knaus <knaus.jeff@gmail.com>