add nix support to app switcher (#1471)
This tweaks the app switcher code a bit to work when using applications
installed via nix on both linux and darwin.

It also removes the `/usr/bin/` assumption for linux executable path,
and uses `which` instead to find the real path instead. Due to pulling
out the path that way now, I had to tweak the regex slightly to get it
to also match if there's only one argument to the app, as it will no
longer be prefixed with a space.