Allow apps to force lang via code.language and add special cased files (#1451)
This makes some tweaks to language modes:

1. It adds a list of special-cased files that are mixed into the
determination of the language, so you can return a language for things
even if the file extension itself doesn't imply the language. The
languages included aren't actually implemented in community, so I can
keep the list empty if preferred, but I figured I'd keep it populated at
first to give an example of why it's useful.
2. Adds a command to show the currently forced language
3. #1256 semi-recently went the route of removing language-specific
tags. This has the side effect (afaict) that an app that doesn't expose
the file extension in its title can no longer force a language
(previously it would just `tag: user.<lang>`). Given we removed the tags
I assume we don't want them back, so this change goes a different route
and adds an action that can be overridden to force a language. It's done
in a way that the language being forced by an app will still be
overridden by a user manually forcing a language with `force <lang>`.
I've included an obsidian skeleton to demonstrate how to use it.
Obsidian is a markdown editor that doesn't include file extensions in
it's title. I took a look at search.talonvoice.com and see that some
people are using tag.markdown (so are their community is not up to date,
or they custom implemented tags again) and some others are
re-implementing markdown commands. Since I tried to use this app
recently, and I don't like forcing a language across my entire system
and have to remember to clear it afterwards, I figured I should add a
way to do it.

I'm open to other ways to do this if there is a better idea.

---------

Co-authored-by: David Vo <auscompgeek@users.noreply.github.com>
Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>