Added contacts feature, documented in core/contacts/README.md. (#1706)
I have been holding onto this one in my fork for a while, yet it's one
of the most useful extensions to community I've written. Long ago I used
to put contacts into vocabulary, but they end up conflicting with other
words. By adding explicit suffixes and a predictable grammar, it is so
much easier to work with them. As a bonus, you never have to remember
email addresses again.

---------

Co-authored-by: Nicholas Riley <com-github@sabi.net>
Co-authored-by: Jeff Knaus <knaus.jeff@gmail.com>