Migrate sql to snippets (#1861)
A possible regression is that the space after the close parentheses for
the with command is lost when migrated to snippets because a space added
to the snippet after the parentheses gets ignored. Based on what I
remember about SQL and some quick research, that does not seem
consequential, but it may be worth checking with people who use SQL more
than I do if this is a regression or not.