Raw Snippet Insert Improvement: Do not move cursor when first stop is at the end (#1908)
The raw snippet insert action will currently move the cursor to try to
put it in the right position even if the last stop is at the end of the
snippet, which should not require any movement. This can cause the
cursor to get moved to the wrong position if the last line only consists
of indentation because the action moves the cursor to the right position
on the target line by first going to the start of the line. This
addresses this issue.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>