Migrate for loop actions to use snippets (#1894)
This migrates all the for loop actions not handled in a separate pull
request. If we are going to migrate the previous state for command, I
recommend having a separate python snippet for "for in" without the
colon for more complex comprehensions. I have no idea what to call it
though.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>