Support customizing indentation with raw text snippet insert (#1907)
Now that we are migrating existing functionality to snippets, I want to
provide settings that can allow customizing how snippets are inserted
without editor support to improve the experience for people using
editors other than VSCode. As the first pull request towards this goal,
I added a setting for deciding if tabs should be converted to spaces and
a setting for determining the number of spaces that should be used per
tab when converting tabs to spaces.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>
Co-authored-by: Nicholas Riley <com-github@sabi.net>