Provide c++ operators definition (#1974)
We can just use the c operators. 

I wanted to work on c++ support one feature at a time as the larger pull
requests adding support for the entire language in one go seem to
consistently lose momentum. It is worth discussing if we should put
support for the language in a separate file. I personally think so to
improve organization. We can use talon constructs and importing to reuse
relevant pieces of the c implementation anyways.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>